requirements.txt
crewai==0.186.1
crewai-tools==0.73.0
python-dotenv==1.1.1
yfinance==0.2.66
requests==2.32.5
beautifulsoup4==4.13.5
pandas==2.3.2
streamlit==1.49.1
tenacity==9.0.0
ratelimit==2.2.1
pyyaml==6.0.2
pydantic==2.10.5
